requirements.txt
kagglehub[pandas-datasets]>=0.3.12
pandas>=2.2.0
numpy>=1.24.0
scikit-learn>=1.4.0
pyyaml>=6.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
lightgbm>=4.3.0
joblib>=1.3.0
pydantic>=2.6.0
mlflow>=2.10.0 